+++ modules/jaspr-code/src/helpers/install_helper.ts
@@ -1,6 +1,6 @@
 import { cwd } from "process";
 import { dartExtensionApi } from "../api";
-import { runCommand } from "./process_helper";
+import { runJaspr } from "./process_helper";
 import * as vs from "vscode";
 import { minimumJasprVersion } from "../constants";
 
@@ -8,20 +8,20 @@ export async function checkJasprVersion(): Promise<string | undefined> {
   let installedVersion = await getInstalledVersion();
 
   try {
+
     if (installedVersion === undefined) {
       const action = await vs.window.showWarningMessage(
-        "jaspr_cli needs to be installed with 'pub global activate jaspr_cli' to use this extension.",
+        "jaspr_cli needs to be installed and available in your PATH to use this extension.",
         "Install Now"
       );
 
       if (action === "Install Now") {
-        await dartExtensionApi.sdk.runPub(cwd(), [
-          "global",
-          "activate",
+        await dartExtensionApi.sdk.runDart(cwd(), [
+          "install",
           "jaspr_cli",
         ]);
       } else {
-        return installedVersion;
+        return undefined;
       }
     } else if (installedVersion < minimumJasprVersion) {
       const action = await vs.window.showWarningMessage(
@@ -30,26 +30,28 @@ export async function checkJasprVersion(): Promise<string | undefined> {
       );
 
       if (action === "Update Now") {
-        await dartExtensionApi.sdk.runPub(cwd(), [
-          "global",
-          "activate",
+        await dartExtensionApi.sdk.runDart(cwd(), [
+          "install",
           "jaspr_cli",
         ]);
       } else {
-        return installedVersion;
+        return undefined;
       }
+    } else {
+      return installedVersion;
     }
 
     installedVersion = await getInstalledVersion();
     if (installedVersion === undefined || installedVersion < minimumJasprVersion) {
-      throw new Error(
+      vs.window.showErrorMessage(
         `Failed to install or update Jaspr CLI. Installed version: ${installedVersion}`
       );
+      return undefined;
     }
   } catch (error) {
     console.error("Error checking Jaspr version:", error);
     vs.window.showErrorMessage(
-      "Failed to install Jaspr CLI. Please try running 'pub global activate jaspr_cli' manually."
+      "Failed to install Jaspr CLI. Please try running 'dart install jaspr_cli' manually."
     );
   }
 
@@ -58,15 +60,8 @@ export async function checkJasprVersion(): Promise<string | undefined> {
 
 async function getInstalledVersion(): Promise<string | undefined> {
   try {
-    const result = await runCommand(["pub", "global", "list"]);
-    const versionMatch = new RegExp(
-      `^jaspr_cli (\\d+\\.\\d+\\.\\d+[\\w.\\-+]*)(?: |$)`,
-      "m"
-    );
-    const match = versionMatch.exec(result);
-    const installedVersion = match ? match[1] : undefined;
-
-    return installedVersion;
+    const result = await runJaspr(["--version"], cwd());
+    return !!result ? result : undefined;
   } catch (error) {
     return undefined;
   }
@@ -77,11 +72,7 @@ export async function checkJasprInstalled(): Promise<boolean> {
   if (v === undefined) {
     return false;
   }
-  if (v < minimumJasprVersion) {
-    await vs.window.showErrorMessage(
-      `Jaspr CLI version is too old. Please update to ${minimumJasprVersion} or later.`
-    );
-    return false;
-  }
   return true;
 }
+
+